refactor: move session related code to session.ts
