+++ src/app/v2/policy-management/manage/actions.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { createPolicyService } from "@/lib/rpc/client";
-import { getTenantAndToken } from "@/lib/rpc/client";
+import { getTenantAndToken } from "@/lib/session/session";
 
 export async function getPolicyGroups() {
   const { accessToken, tenant } = await getTenantAndToken();
