+++ src/app/v2/policy/new-policy/actions.ts
@@ -1,6 +1,7 @@
 "use server";
 
-import { createPolicyService, getTenantAndToken } from "@/lib/rpc/client";
+import { createPolicyService } from "@/lib/rpc/client";
+import { getTenantAndToken } from "@/lib/session/session";
 import { CreatePolicyRequest } from "@buf/safedep_api.bufbuild_es/safedep/services/controltower/v1/policy_pb";
 
 export async function createPolicy(policy: CreatePolicyRequest) {
